requirements.txt
aiohttp==3.6.3
async-timeout==3.0.1
attrs==20.2.0
beautifulsoup4==4.9.3
cachetools==4.1.1
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
click==7.1.2
dialogflow==1.1.0
discord.py @ git+https://github.com/Gorialis/discord.py@17ca1ca52ecde380e78bd9d518564e4f25ab7924
enum34==1.1.10
ffmpeg==1.4
fuzzywuzzy==0.18.0
google-api-core==1.23.0
google-auth==1.22.1
googleapis-common-protos==1.52.0
grpcio==1.33.1
gTTS==2.1.1
gTTS-token==1.1.3
idna==2.10
multidict==4.7.6
numpy==1.19.2
protobuf==3.13.0
pvporcupine==1.8.7
pvporcupinedemo==1.8.7
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyAudio==0.2.11
pycparser==2.20
pydub==0.24.1
PyNaCl==1.3.0
python-Levenshtein==0.12.0
pytz==2020.1
requests==2.24.0
rsa==4.6
simpleaudio==1.0.4
six==1.15.0
SoundFile==0.10.3.post1
soupsieve==2.0.1
SpeechRecognition==3.8.1
urllib3==1.25.11
yarl==1.5.1
